djangoProject/requirements.txt
Django==4.1.1
djangorestframework==3.14.0
python-dotenv==0.20.0
psycopg2-binary==2.9.3
django-cors-headers==3.11.0